// assets/script/model/DeployGroup.ts
export enum DeployGroup {

    /**
     * 障碍物
     */
    road = 0,
    /**
     * 场景
     */
    beautifyScene = 1,
    /**
     * 道具
     */
    buff = 2,
    /**
     * 敌人
     */
    enemy = 3,
    /**
     * 起点
     */
    start = 4,
    /**
     * 终点
     */
    end = 5,
    /**
     * 陷阱
     */
    trap = 6,
    /**
     * 移动平台
     */
    side = 7
}